projects/cursor-audio-agent/pyproject.toml
[project]
name = "cursor-audio-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.5.4",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "pocketsphinx>=5.0.4",
    "pyaudio>=0.2.14",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.1",
    "speechrecognition>=3.14.3",
]
